+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.8.0] - 2021-05-14
+### Added
+- Enable adding Trace-Id header to response
+### Fixed
+- Use 'rewritePath' and 'rewriteMethod' rule default value
+
 ## [1.7.0] - 2021-02-26
 ### Added
 - support for dynamic port registration in Consul